api/requirements.txt
python-dotenv
werkzeug==2.1.2
Flask~=2.1.2
flask-restx==0.5.1
click<8.1.0
git+https://github.com/blastbeng/ChatterBot
#git+https://github.com/blastbeng/chatterbot-corpus
chatterbot-corpus
spacy
pyyaml==6.0.0
ruamel.yaml==0.17.21
itsdangerous
Jinja2
MarkupSafe
uWSGI
wikipedia
gtts
git+https://github.com/kinshuk-h/pytube
pysqlite3
redis==4.3.4
Flask-Caching==2.0.1
translate
Flask-APScheduler
Tweepy
faker
praw
pyenchant
markovipy
fakeyou
sqlitedict
Flask-Limiter
uwsgi-tools